requirements.txt
# Claims Processing System - Python Dependencies
# Source: Design Document - Technical Requirements
# Verified: 2025-12-18

# =============================================================================
# Core Framework
# =============================================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# =============================================================================
# Database
# =============================================================================
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# =============================================================================
# Caching & Messaging
# =============================================================================
redis>=5.0.0
aioredis>=2.0.0

# =============================================================================
# Authentication & Security
# =============================================================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0
python-multipart>=0.0.6

# =============================================================================
# HTTP & Networking
# =============================================================================
httpx>=0.26.0
aiohttp>=3.9.0
requests>=2.31.0

# =============================================================================
# Data Processing
# =============================================================================
pandas>=2.1.0
numpy>=1.26.0
openpyxl>=3.1.0
python-dateutil>=2.8.0

# =============================================================================
# AI/ML
# =============================================================================
scikit-learn>=1.4.0
sentence-transformers>=2.2.0
# torch>=2.1.0  # Optional: for GPU acceleration

# =============================================================================
# Document Processing
# =============================================================================
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.2.0
PyMuPDF>=1.24.0
python-magic>=0.4.27

# =============================================================================
# Search Engine
# Source: Design Document 04_validation_engine_comprehensive_design.md
# =============================================================================
typesense>=0.21.0

# =============================================================================
# LLM Gateway
# Source: Design Document 04_validation_engine_comprehensive_design.md
# =============================================================================
litellm>=1.50.0

# =============================================================================
# Monitoring & Observability
# =============================================================================
prometheus-client>=0.19.0
structlog>=24.1.0
python-json-logger>=2.0.7

# =============================================================================
# UI Dashboard
# =============================================================================
streamlit>=1.30.0
plotly>=5.18.0

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-rerunfailures>=13.0
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
isort>=5.13.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0
PyYAML>=6.0.0
tenacity>=8.2.0
